LayoutTests/platform/ios-sim/css3/device-adapt/opera/constrain-013-expected.txt


FAIL CSS Test: @viewport constrained - max-zoom affecting width. assert_equals: expected 1600 but got 800(stack: assert@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:705:15
file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-013.xhtml:69:22
step@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:1186:30
onload@file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-013.xhtml:68:14)
